Created Console Blackjack Game using Java
This is a group assignment for my Software Engineering class.  This
initial code/commit is entirely mine but assuming my group decides to
use this code the plan is to add a GUI(Graphical User Interface).
However this isn't the only set of code there is so there is the
possibility they will reject this code in favor of the other code which
is entirely fine of course.  I loved working on this.  My code ended up
spaghetti-coded and it's ugly but it worked and it was fun to break down
the classes in my head.